requirements.txt
mcp[cli]
httpx
requests
beautifulsoup4
PyGithub
web3
python-dotenv
tweepy
